Mucho avance en pong. Movimiento, keypressed y keyreleased y varios arreglos

+++ Kernel/include/time.h
@@ -1,8 +1,10 @@
 #ifndef TIME_H
 #define TIME_H
 
+#include <stdint.h>
+
 void timer_handler();
-int ticks_elapsed();
-int seconds_elapsed();
+uint64_t ticks_elapsed();
+uint64_t seconds_elapsed();
 
 #endif
